Fixed a bug when sizeof(size_t) != sizeof(int).
The PyArg_ParseTuple "s#" formatter requires an int (and _not_ a
size_t) for the string length.  On 64-bit platforms the higher bits
were often left as garbage, leading to an out of memory condition
later in libtre.  Then, a second yet unfixed bug would lead to a
crash.  This fix at least takes care of the first bug.

darcs-hash:20090904171748-ced27-681b87142937f3f01da7fb60d7787e254d000038.gz

+++ python/tre-python.c
@@ -471,7 +471,7 @@ PyTre_ncompile(PyObject *self, PyObject *args)
 {
   TrePatternObject *rv;
   char *pattern;
-  size_t pattlen;
+  int pattlen;
   int cflags = 0;
   int rc;
